requirements.txt
fastapi
uvicorn[standard]
chromadb
sentence-transformers
pymupdf
python-docx
pandas
pillow
python-multipart
numpy
